+++ code/proxy.ts
@@ -27,8 +27,10 @@ export function proxy(request: NextRequest) {
     // Check if there is any supported locale in the pathname
     const { pathname } = request.nextUrl
 
-    // Ignores all manifests
-    if (pathname.includes("/manifest_") && pathname.endsWith(".json")) {
+    // Ignores all public directory contents and manifests
+    if (pathname.includes("config") || pathname.includes("images") || pathname.includes("optional-pages") ||
+        (pathname.includes("/manifest_") && pathname.endsWith(".json")) ||
+        pathname.includes("/sw.js") || pathname.includes("/serwist/")) {
         return;
     }
     const pathnameHasLocale = supportedLocales.some(