+++ code/ui/text/status/status.tsx
@@ -33,6 +33,8 @@ export function getTranslatedStatusLabel(
       return dict.title.accrued;
     case "issue":
       return dict.title.issue;
+    case RegistryStatusMap.VOIDED:
+      return dict.title.voided;
     case "rescinded":
       return dict.title.rescinded;
     case "terminated":
@@ -77,6 +79,11 @@ export default function StatusComponent(props: Readonly<StatusComponentProps>) {
       statusTextColor = "var(--success-foreground)";
       statusBackgroundColor = "var(--success-background)";
       statusBorderColor = "var(--success-border)";
+      break;
+    case dict.title.voided.toLowerCase():
+      statusTextColor = "var(--void-foreground)";
+      statusBackgroundColor = "var(--void-background)";
+      statusBorderColor = "var(--void-border)";
       break;
     case dict.title.rescinded.toLowerCase():
     case dict.title.terminated.toLowerCase():